:recycle: 1ファイル1クラスのルールに準拠
