+++ AvatarManager.Core/Infrastructures/Database/ApplicationDbContextFactory.cs
@@ -1,0 +1,19 @@
+﻿using Microsoft.Data.Sqlite;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Design;
+
+namespace AvatarManager.Core.Infrastructures.Database;
+
+public class ApplicationDbContextFactory : IDesignTimeDbContextFactory<ApplicationDbContext>
+{
+    public ApplicationDbContext CreateDbContext(string[] args)
+    {
+        var connection = new SqliteConnection("Data Source=AvatarManager.db");
+        connection.Open();
+        var options = new DbContextOptionsBuilder<ApplicationDbContext>()
+            .UseSqlite(connection)
+            .Options;
+        return new ApplicationDbContext(options);
+    }
+}
+
